Orbslam experiment config updates
